# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Testing Swagger APIs.",
        "title": "Admin Course APIs",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "45.141.79.220",
    "basePath": "/v1",
    "paths": {
        "/banner": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "all data banner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banner"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/banner.DataBannerDTO"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "create banner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banner"
                ],
                "parameters": [
                    {
                        "description": "Create JSON",
                        "name": "banner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/banner.AddBannerDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/banner.GetBannerDTO"
                        }
                    }
                }
            }
        },
        "/banner/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "one data banner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banner"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "banner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/banner.GetBannerDTO"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update banner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banner"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "banner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create JSON",
                        "name": "banner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/banner.AddBannerDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "delete banner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banner"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "banner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/news": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "all data news",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banner"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/news.DataNewsDTO"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "create banner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "parameters": [
                    {
                        "description": "Create JSON",
                        "name": "news",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/news.AddNewsDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/news.GetNewsDTO"
                        }
                    }
                }
            }
        },
        "/news/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "one data news",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "news id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/news.GetNewsDTO"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update news",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "news id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create JSON",
                        "name": "news",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/news.AddNewsDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/news.GetNewsDTO"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "delete banner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banner"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "news id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "banner.AddBannerDTO": {
            "type": "object",
            "properties": {
                "imagePath": {
                    "type": "string"
                }
            }
        },
        "banner.DataBannerDTO": {
            "type": "object",
            "properties": {
                "banners": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/banner.GetBannerDTO"
                    }
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "banner.GetBannerDTO": {
            "type": "object",
            "properties": {
                "Id": {
                    "type": "integer"
                },
                "imagePath": {
                    "type": "string"
                }
            }
        },
...
        "news.AddNewsDTO": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "imagePath": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "news.DataNewsDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "news": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/news.GetNewsDTO"
                    }
                }
            }
        },
        "news.GetNewsDTO": {
            "type": "object",
            "properties": {
                "Id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "imagePath": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}